Moved DirectBackupEntryProcessor to its own impl class and made it IdentifiedDataSerializable.
Moved EntryProcessorUtil to the correct package.
