+++ hazelcast/src/main/java/com/hazelcast/map/EntryProcessorUtil.java
@@ -14,30 +14,14 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map.listener;
+package com.hazelcast.map;
 
-import com.hazelcast.map.EntryProcessor;
-
-import java.util.Map;
+import com.hazelcast.map.impl.DirectBackupEntryProcessor;
 
 /**
  * Utility class for {@link EntryProcessor}.
  */
 public final class EntryProcessorUtil {
-
-    private static final DirectBackupEntryProcessor DIRECT_BACKUP_ENTRY_PROCESSOR = new DirectBackupEntryProcessor();
-
-    private static final class DirectBackupEntryProcessor<K, V, R> implements EntryProcessor<Object, Object, Object> {
-
-        @Override
-        public Object process(Map.Entry<Object, Object> entry) {
-            return null;
-        }
-
-        private Object readResolve() {
-            return DIRECT_BACKUP_ENTRY_PROCESSOR;
-        }
-    }
 
     private EntryProcessorUtil() {
     }
@@ -50,10 +34,11 @@ public final class EntryProcessorUtil {
      * This may be useful in cases when loading the backup entries and/or applying
      * the backup entry processor is costly and direct copy of the modifications
      * is more efficient.
+     *
      * @return the direct backup processor singleton.
      */
     @SuppressWarnings("unchecked")
     public static <K, V, R> EntryProcessor<K, V, R> directBackupProcessor() {
-        return (EntryProcessor<K, V, R>) DIRECT_BACKUP_ENTRY_PROCESSOR;
+        return (EntryProcessor<K, V, R>) DirectBackupEntryProcessor.DIRECT_BACKUP_PROCESSOR;
     }
 }
