+++ hazelcast/src/main/java/com/hazelcast/map/impl/DirectBackupEntryProcessor.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2008-2020, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hazelcast.map.impl;
+
+import com.hazelcast.map.EntryProcessor;
+import com.hazelcast.nio.ObjectDataInput;
+import com.hazelcast.nio.ObjectDataOutput;
+import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
+
+import java.util.Map;
+
+public class DirectBackupEntryProcessor implements EntryProcessor<Object, Object, Object>, IdentifiedDataSerializable {
+
+    public static final DirectBackupEntryProcessor DIRECT_BACKUP_PROCESSOR = new DirectBackupEntryProcessor();
+
+    public DirectBackupEntryProcessor() {
+    }
+
+    @Override
+    public Object process(Map.Entry<Object, Object> entry) {
+        return null;
+    }
+
+    @Override
+    public int getFactoryId() {
+        return MapDataSerializerHook.F_ID;
+    }
+
+    @Override
+    public int getClassId() {
+        return MapDataSerializerHook.DIRECT_BACKUP_ENTRY_PROCESSOR;
+    }
+
+    @Override
+    public void writeData(ObjectDataOutput out) {
+    }
+
+    @Override
+    public void readData(ObjectDataInput in) {
+    }
+}
