+++ hazelcast/src/main/java/com/hazelcast/map/impl/MapDataSerializerHook.java
@@ -312,8 +312,9 @@ public final class MapDataSerializerHook implements DataSerializerHook {
     public static final int PUT_IF_ABSENT_WITH_EXPIRY = 146;
     public static final int PUT_TRANSIENT_BACKUP = 147;
     public static final int PUT_ALL_DELETE_ALL_BACKUP = 148;
+    public static final int DIRECT_BACKUP_ENTRY_PROCESSOR = 149;
 
-    private static final int LEN = PUT_ALL_DELETE_ALL_BACKUP + 1;
+    private static final int LEN = DIRECT_BACKUP_ENTRY_PROCESSOR + 1;
 
     @Override
     public int getFactoryId() {
@@ -470,6 +471,7 @@ public final class MapDataSerializerHook implements DataSerializerHook {
         constructors[PUT_IF_ABSENT_WITH_EXPIRY] = arg -> new PutIfAbsentWithExpiryOperation();
         constructors[PUT_TRANSIENT_BACKUP] = arg -> new PutTransientBackupOperation();
         constructors[PUT_ALL_DELETE_ALL_BACKUP] = arg -> new PutAllDeleteAllBackupOperation();
+        constructors[DIRECT_BACKUP_ENTRY_PROCESSOR] = arg -> DirectBackupEntryProcessor.DIRECT_BACKUP_PROCESSOR;
 
         return new ArrayDataSerializableFactory(constructors);
     }
